Rename project to learningRedux
