+++ android/app/src/main/java/com/learningredux/MainActivity.java
@@ -1,4 +1,4 @@
-package com.react036;
+package com.learningredux;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "react036";
+        return "learningRedux";
     }
 }
