+++ android/gradle.properties
@@ -18,3 +18,6 @@
 # org.gradle.parallel=true
 
 android.useDeprecatedNdk=true
+
+org.gradle.daemon=true
+org.gradle.parallel=true
