+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'react036'
+rootProject.name = 'learningRedux'
 
 include ':app'
