+++ ios/learningRedux.xcodeproj/xcshareddata/xcschemes/learningRedux.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "react036.app"
-               BlueprintName = "react036"
-               ReferencedContainer = "container:react036.xcodeproj">
+               BuildableName = "learningRedux.app"
+               BlueprintName = "learningRedux"
+               ReferencedContainer = "container:learningRedux.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "react036Tests.xctest"
-               BlueprintName = "react036Tests"
-               ReferencedContainer = "container:react036.xcodeproj">
+               BuildableName = "learningReduxTests.xctest"
+               BlueprintName = "learningReduxTests"
+               ReferencedContainer = "container:learningRedux.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -47,9 +47,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "react036Tests.xctest"
-               BlueprintName = "react036Tests"
-               ReferencedContainer = "container:react036.xcodeproj">
+               BuildableName = "learningReduxTests.xctest"
+               BlueprintName = "learningReduxTests"
+               ReferencedContainer = "container:learningRedux.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -57,9 +57,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "react036.app"
-            BlueprintName = "react036"
-            ReferencedContainer = "container:react036.xcodeproj">
+            BuildableName = "learningRedux.app"
+            BlueprintName = "learningRedux"
+            ReferencedContainer = "container:learningRedux.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
    </TestAction>
@@ -77,9 +77,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "react036.app"
-            BlueprintName = "react036"
-            ReferencedContainer = "container:react036.xcodeproj">
+            BuildableName = "learningRedux.app"
+            BlueprintName = "learningRedux"
+            ReferencedContainer = "container:learningRedux.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -96,9 +96,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "react036.app"
-            BlueprintName = "react036"
-            ReferencedContainer = "container:react036.xcodeproj">
+            BuildableName = "learningRedux.app"
+            BlueprintName = "learningRedux"
+            ReferencedContainer = "container:learningRedux.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
