+++ ios/learningRedux/Images.xcassets/AppIcon.appiconset/Contents.json
