+++ ios/learningRedux/Info.plist
