+++ ios/learningReduxTests/Info.plist
