+++ ios/learningReduxTests/learningReduxTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface react036Tests : XCTestCase
+@interface learningReduxTests : XCTestCase
 
 @end
 
-@implementation react036Tests
+@implementation learningReduxTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
