+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react036",
+  "name": "learningRedux",
   "version": "0.0.1",
   "private": true,
   "scripts": {
